analysis-service/requirements.txt
boto3==1.34.0
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
aiofiles==23.2.1
psutil==5.9.6
ffmpeg-python==0.2.0
